タスク移動時にタスクが初期化される不具合修正
